app/requirements.txt
flask
flask-sqlalchemy
uwsgi
requests
ruamel.yaml
RPi.GPIO

# dev:
flake8
